requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
django-dotenv==1.4.2
django-environ==0.11.2
django-jazzmin==3.0.0
django-modeltranslation==0.19.5
django-phonenumber-field==8.0.0
django-rosetta==0.10.0
djangorestframework==3.15.2
drf-yasg==1.21.7
environ==1.0
idna==3.7
inflection==0.5.1
packaging==24.1
phonenumbers==8.13.40
polib==1.2.0
psycopg==3.2.1
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
